move: headline news ticker data & content load
